Compiler changes for builtin native Object type
This is just a test to see how implementing a SP native version of a Handle would work, in this case a Object, or could be referred to as a native object as it can refer to a type in native code.
For example a native object could wrap a std::vector.